requirements.txt
graphviz>=0.20.1
Pillow>=10.0.0
networkx>=3.0
numpy>=1.24.0
matplotlib>=3.7.0